atualizando

+++ Imc/src/imc/Pessoa.java
@@ -70,7 +70,7 @@ public class Pessoa {
         return altura;
     }
 
-    public void setAltura(int altura) {
+    public void setAltura(double altura) {
         this.altura = altura;
     }
